Add most functions. Just test part need to be done.

+++ Cenux.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 
 __all__ = ['Cenux_datapoint', 'Cenux_question',
            'Cenux_user', 'Cenux_competition']
@@ -30,8 +31,8 @@ class Cenux_datapoint(object):
         if not (os.path.exists(outputfile) and os.path.isfile(outputfile)):
             raise ValueError(
                 "Output file {0} is not valid.".format(outputfile))
-        self.inputfile = str(inputfile)
-        self.outputfile = str(outputfile)
+        self.inputfile = os.path.abspath(inputfile)
+        self.outputfile = os.path.abspath(outputfile)
         self.timelimit = float(timelimit)
         self.memorylimit = float(memorylimit)
         if not score == None:
@@ -69,11 +70,16 @@ class Cenux_datapoint(object):
 
     def set_inputfile(self, inputfile, **kwargs):
         '''Set the input file path for data point'''
-        self.inputfile = str(inputfile)
+        if not (os.path.exists(inputfile) and os.path.isfile(inputfile)):
+            raise ValueError("Input file {0} is not valid.".format(inputfile))
+        self.inputfile = os.path.abspath(inputfile)
 
     def set_outputfile(self, outputfile, **kwargs):
         '''Set the output file path for data point'''
-        self.outputfile = str(outputfile)
+        if not (os.path.exists(outputfile) and os.path.isfile(outputfile)):
+            raise ValueError(
+                "Output file {0} is not valid.".format(outputfile))
+        self.outputfile = os.path.abspath(outputfile)
 
     def set_score(self, score, **kwargs):
         '''Set the score for data point'''
@@ -87,25 +93,29 @@ class Cenux_question(object):
     """This is a question class for Cenux."""
 
     def __init__(self, name=None, sourcefilename=None, timelimit=1.,
-                 memorylimit=128., datapoint=[], **kwargs):
+                 memorylimit=128., datapoint=[], copypath=None, **kwargs):
         if not sourcefilename:
             sourcefilename = name
         self.name = str(name)
         self.sourcefilename = str(sourcefilename)
         self.timelimit = float(timelimit)
         self.memorylimit = float(memorylimit)
+        self.copypath = os.path.abspath(copypath)
         self.__kwargs = kwargs
-        if type(datapoint) == list and all([isinstance(i, Cenux_datapoint) for i in datapoint]):
+        if (type(datapoint) == list and
+                all([isinstance(i, Cenux_datapoint) for i in datapoint])):
             self.datapoint = datapoint
         else:
             self.datapoint = []
             print "*** Could not read datapoint. Datapoint set to empty list."
 
     def __str__(self):
-        return 'Cenux Datapoint ' + self.name + '({0} datapoints)'.format(len(self.datapoint))
+        return 'Cenux Datapoint ' + self.name + \
+            '({0} datapoints)'.format(len(self.datapoint))
 
     def __repr__(self):
-        return 'Cenux Datapoint ' + self.name + '({0} datapoints)'.format(len(self.datapoint))
+        return 'Cenux Datapoint ' + self.name + \
+            '({0} datapoints)'.format(len(self.datapoint))
 
     def __getitem__(self, index):
         return self.datapoint[index]
@@ -123,7 +133,8 @@ class Cenux_question(object):
                 (all([self[i] == other[i]
                       for i in range(len(self.datapoint))])))
 
-    def add_datapoint(self, inputfile, outputfile, score=None, auto=1, **kwargs):
+    def add_datapoint(self, inputfile, outputfile, score=None, auto=1,
+                      **kwargs):
         '''Add datapoint to the question.
         inputfile: inpuut file path
         outputfile: output file path
@@ -134,12 +145,42 @@ class Cenux_question(object):
         auto: add similar datapoints automaticly, default as 1
         '''
 
+        if not os.path.isfile(inputfile) or not os.path.isfile(outputfile):
+            raise ValueError("The files couldn't be recongnized.")
         info = kwargs.pop("info", 1)
         timelimit = kwargs.pop("timelimit", self.timelimit)
         memorylimit = kwargs.pop("memorylimit", self.memorylimit)
+        kwargs.pop("index", None)
+        if self.copypath:
+            if (os.path.abspath(inputfile) ==
+                    os.path.join(self.copypath, os.path.split(inputfile)[1]) and
+                    os.path.abspath(outputfile) ==
+                    os.path.join(self.copypath, os.path.split(outputfile)[1])):
+                inputfile = os.path.abspath(inputfile)
+                outputfile = os.path.abspath(outputfile)
+            elif (os.path.exists(
+                    os.path.join(self.copypath, os.path.split(inputfile)[1])) or
+                    os.path.exists(
+                    os.path.join(self.copypath, os.path.split(outputfile)[1]))):
+                raise ValueError("Same name file exists.")
+            else:
+                shutil.copyfile(inputfile,
+                                os.path.join(self.copypath, os.path.split(inputfile)[1]))
+                shutil.copyfile(outputfile,
+                                os.path.join(self.copypath, os.path.split(outputfile)[1]))
+                inputfile = os.path.join(self.copypath,
+                                         os.path.split(inputfile)[1])
+                outputfile = os.path.join(self.copypath,
+                                          os.path.split(inputfile)[1])
         self.datapoint = self.datapoint + \
-            [Cenux_datapoint(inputfile, outputfile, timelimit=timelimit, memorylimit=memorylimit,
-                             score=score, info=info, index=len(self.datapoint), **kwargs)]
+            [Cenux_datapoint(inputfile,
+                             outputfile,
+                             timelimit=timelimit,
+                             memorylimit=memorylimit,
+                             score=score,
+                             info=info,
+                             index=len(self.datapoint),
+                             **kwargs)]
         if auto:
             print "\n## Add datapoint autometicly."
             inpath, inname = os.path.split(inputfile)
@@ -150,32 +191,58 @@ class Cenux_question(object):
             outmatch = rr.match(outname)
             if not(inmatch and outmatch):
                 print "*** Couldn't recongnize the file name. Only 1 datapoint added."
-            elif inmatch.span() != (0, len(inname)) or outmatch.span() != (0, len(outname)):
+            elif (inmatch.span() != (0, len(inname))
+                    or outmatch.span() != (0, len(outname))):
                 print "*** Couldn't recongnize the file name. Only 1 datapoint added."
             elif inmatch.groups()[1] != outmatch.groups()[1]:
                 print "*** Couldn't recongnize the file name. Only 1 datapoint added."
             else:
                 informat = inmatch.groups()
                 outformat = outmatch.groups()
-                intemp = [j for j in [i for i in os.listdir(inpath) if rr.match(i)] if rr.match(
-                    j).span() == (0, len(j)) and rr.match(j).groups()[::2] == informat[::2]]
-                outtemp = [j for j in [i for i in os.listdir(outpath) if rr.match(i)] if rr.match(
-                    j).span() == (0, len(j)) and rr.match(j).groups()[::2] == outformat[::2]]
+                intemp = [j for j in [i for i in os.listdir(inpath) if rr.match(i)]
+                          if rr.match(j).span() == (0, len(j)) and rr.match(j).groups()[::2] == informat[::2]]
+                outtemp = [j for j in [i for i in os.listdir(outpath) if rr.match(i)]
+                           if rr.match(j).span() == (0, len(j)) and rr.match(j).groups()[::2] == outformat[::2]]
                 informat1 = list(informat)
                 outformat1 = list(outformat)
-                for i in sorted(list(set([rr.match(i).groups()[1] for i in intemp]).intersection(set([rr.match(i).groups()[1] for i in outtemp]))), cmp=lambda x, y: cmp(int(x), int(y))):
+                for i in sorted(list(set([rr.match(i).groups()[1] for i in intemp])
+                                     .intersection(set([rr.match(i).groups()[1] for i in outtemp]))),
+                                cmp=lambda x, y: cmp(int(x), int(y))):
                     if i == informat[1]:
                         continue
                     informat1[1] = i
                     outformat1[1] = i
-                    self.datapoint = self.datapoint + [Cenux_datapoint(os.path.join(inpath, ''.join(informat1)),
-                                                                       os.path.join(
-                        outpath, ''.join(outformat1)),
-                        timelimit=timelimit,
-                        memorylimit=memorylimit,
-                        score=score,
-                        info=info,
-                        index=len(self.datapoint), **kwargs)]
+                    iii = os.path.join(inpath, ''.join(informat1))
+                    ooo = os.path.join(outpath, ''.join(outformat1))
+                    if self.copypath:
+                        if (os.path.abspath(iii) == os.path.join(self.copypath, os.path.split(iii)[1])
+                                and os.path.abspath(ooo) == os.path.join(self.copypath, os.path.split(ooo)[1])):
+                            iii = os.path.abspath(iii)
+                            ooo = os.path.abspath(ooo)
+                        elif (os.path.exists(os.path.join(self.copypath, os.path.split(iii)[1]))
+                                or os.path.exists(os.path.join(self.copypath, os.path.split(ooo)[1]))):
+                            print "## File {0} or {1} exists. Ignored.".format(iii, ooo)
+                            iii = ooo = None
+                        else:
+                            shutil.copyfile(
+                                iii, os.path.join(self.copypath, os.path.split(iii)[1]))
+                            shutil.copyfile(
+                                ooo, os.path.join(self.copypath, os.path.split(ooo)[1]))
+                            iii = os.path.join(
+                                self.copypath, os.path.split(iii)[1])
+                            ooo = os.path.join(
+                                self.copypath, os.path.split(ooo)[1])
+                    if iii and ooo:
+                        if not any([(tempdp.inputfile == iii and tempdp.outputfile == ooo) for tempdp in self.datapoint]):
+                            self.datapoint = self.datapoint + \
+                                [Cenux_datapoint(iii,
+                                                 ooo,
+                                                 timelimit=timelimit,
+                                                 memorylimit=memorylimit,
+                                                 score=score,
+                                                 info=info,
+                                                 index=len(self.datapoint),
+                                                 **kwargs)]
 
     def set_timelimit(self, timelimit=1., **kwargs):
         '''Set the time limit for all data point in the question.(second)
@@ -240,6 +307,22 @@ class Cenux_question(object):
         for i in range(len(self.datapoint)):
             self.datapoint[i].index = i
 
+    def collect_to_path(self, path, **kwargs):
+        if not (os.path.exists(path) and os.path.isdir(path)):
+            raise ValueError("The path is invalid.")
+        path = os.path.abspath(path)
+        for i in range(len(self.datapoint)):
+            filename = os.path.split(self.datapoint[i].inputfile)[1]
+            shutil.copyfile(
+                self.datapoint[i].inputfile, os.path.join(path, filename))
+            self.datapoint[i].inputfile = os.path.abspath(
+                os.path.join(path, filename))
+            filename = os.path.split(self.datapoint[i].outputfile)[1]
+            shutil.copyfile(
+                self.datapoint[i].outputfile, os.path.join(path, filename))
+            self.datapoint[i].outputfile = os.path.abspath(
+                os.path.join(path, filename))
+
 
 class Cenux_user(object):
 
@@ -255,15 +338,6 @@ class Cenux_user(object):
 
     def __repr__(self):
         return 'Cenux contestant ' + self.name
-
-    def add_code(self, path, **kwargs):
-        if path == None:
-            self.code = self.code + [None]
-        else:
-            if not(os.path.exists(path) and os.path.isfile(path)):
-                raise ValueError("The path for the code is invalid.")
-            self.code = self.code + [path]
-        self.result = self.result + [None]
 
     def clear_code(self, **kwargs):
         self.code = []
@@ -282,6 +356,16 @@ class Cenux_competition(object):
         if not path:
             path = os.path.abspath(os.getcwd())
         self.abspath = path
+        if ((os.path.exists(os.path.join(self.abspath, "data"))
+                and os.path.isfile(os.path.join(self.abspath, "data"))) or
+            (os.path.exists(os.path.join(self.abspath, "src"))
+                and os.path.isfile(os.path.join(self.abspath, "src")))):
+            raise ValueError(
+                "data or src folder are file, please delete them and try again.")
+        if not os.path.exists(os.path.join(self.abspath, 'data')):
+            os.mkdir(os.path.join(self.abspath, 'data'))
+        if not os.path.exists(os.path.join(self.abspath, 'src')):
+            os.mkdir(os.path.join(self.abspath, 'src'))
         self.__fullfilepath = os.path.join(self.abspath, name + '.Cenux')
         f = open(self.__fullfilepath, 'w')
         f.close()
@@ -289,10 +373,12 @@ class Cenux_competition(object):
         self.user = []
 
     def __str__(self):
-        return 'Cenux Competiiton ' + self.name + '({0} qustions)'.format(len(self.question))
+        return 'Cenux Competiiton ' + self.name \
+            + '({0} qustions)'.format(len(self.question))
 
     def __repr__(self):
-        return 'Cenux Competiiton ' + self.name + '({0} qustions)'.format(len(self.question))
+        return 'Cenux Competiiton ' + self.name \
+            + '({0} qustions)'.format(len(self.question))
 
     def __getitem__(self, index):
         return self.question[index]
@@ -314,17 +400,21 @@ class Cenux_competition(object):
         f.write(a)
         f.close()
 
-    def add_qustion(self, name, sourcefilename=None, timelimit=1., memorylimit=128., datapoint=[], inputfile=None, outputfile=None, score=None, **kwargs):
+    def add_qustion(self, name, sourcefilename=None, timelimit=1.,
+                    memorylimit=128., datapoint=[], inputfile=None,
+                    outputfile=None, score=None, **kwargs):
         if not sourcefilename:
             sourcefilename = name
-        self.question = self.question + [Cenux_question(name=name,
-                                                        sourcefilename=sourcefilename,
-                                                        timelimit=timelimit,
-                                                        memorylimit=memorylimit,
-                                                        datapoint=datapoint)]
+        self.question = self.question + \
+            [Cenux_question(name=name,
+                            sourcefilename=sourcefilename,
+                            timelimit=timelimit,
+                            memorylimit=memorylimit,
+                            datapoint=datapoint,
+                            copypath=os.path.join(self.abspath, 'data'))]
         if inputfile and outputfile:
             self.question[-1].add_datapoint(inputfile,
-                                            outputfile, score, 1, info=1)
+                                            outputfile, score, auto=1, info=1)
 
     def add_user(self, name=None, path=None, **kwargs):
         '''Add a user to the competition.
@@ -333,9 +423,12 @@ class Cenux_competition(object):
         if not name:
             name = str(len(self.user) + 1)
         self.user = self.user + [Cenux_user(name)]
+        os.mkdir(os.path.join(self.abspath, 'src', name))
         print '## User {0} has been added.'.format(name)
         auto = kwargs.pop('auto', None)
         exp = kwargs.pop(exp, ['.c', '.cpp', '.c++', '.pas'])
+        self.user[-1].code = [None] * len(self.question)
+        self.user[-1].result = [None] * len(self.question)
         if auto and not path:
             path = os.getcwd()
         if path:
@@ -344,19 +437,25 @@ class Cenux_competition(object):
             else:
                 self.collect_code(len(user) - 1, path, exp=exp)
 
-    def collect_code(self, index, path, exp=['.c', '.cpp', '.c++', '.pas'] ** kwargs):
+    def collect_code(self, index, path, exp=['.c', '.cpp', '.c++', '.pas'],
+                     ** kwargs):
         try:
             self.user[index]
         except:
             raise ValueError("The index couldn't be recognized.")
-        for i in self.question:
-            temp = __searchfile(path, i.sourcefilename, exp)
+        if self.user[index].code == []:
+            self.user[index].code = [None] * len(self.question)
+            self.user[index].result = [None] * len(self.question)
+        for i in range(len(self.question)):
+            temp = __searchfile(path, self.question[i].sourcefilename, exp)
             if len(temp) == 0:
-                self.user[i].code += [None]
+                pass
             elif len(temp) == 1:
-                # self.user[i].code+=[None]
-                self.add_code(os.path.relpath(temp[0], self.abspath))
+                shutil.copyfile(temp[0],
+                                os.path.join(self.abspath, 'src', os.path.split(temp[0])[1]))
+                self.user[index].code[i] = os.path.join(
+                    self.abspath, 'src', os.path.split(temp[0])[1])
+                self.user[index].result[i] = None
             else:
-                print '## There are more than one file for question {0}'.format(i.name)
+                print '## There are more than one file for question {0}'.format(self.question[i].name)
                 print '## You could set exp option.'
-                for j in self.user[i].code:
